+++ tests/testthat/test-ratatouille.R
@@ -1,6 +1,6 @@
 # Don't actually get all records for tests: mock listing function to only return
 # 50 random records instead of all
-random_object_ids <- sample(list_object_ids(), size = 50)
+random_object_ids <- sample(list_object_ids("rato"), size = 50)
 
 local_mocked_bindings(
   list_object_ids = function(...) random_object_ids
@@ -26,7 +26,7 @@ test_that("ratatouille returns error on invalid source", {
 
 test_that("ratatouille can pass on arguments to its internal helpers", {
   expect_s3_class(
-    ratatouille(batch_size = 1),
+    ratatouille("rato", batch_size = 1),
     "data.frame"
   )
 })