+++ tests/testthat/test-utils.R
@@ -26,3 +26,101 @@ test_that("as_datetime() succesfully converts a few known examples", {
     )
   )
 })
+
+test_that("get_api_domain() returns correct API base urls", {
+  expect_identical(
+    get_api_domain("rato"),
+    "https://gis.oost-vlaanderen.be"
+  )
+  expect_identical(
+    get_api_domain("wfl"),
+    "https://gwadmin.west-vlaanderen.be"
+  )
+})
+
+test_that("get_api_domain() returns error on multiple sources", {
+  expect_error(
+    get_api_domain(c("rato", "wfl")),
+    class = "rata_multiple_sources"
+  )
+})
+
+test_that("check_credentials() returns an error if credentials are unset", {
+  withr::with_envvar(
+    new = c("RATO_PWD" = ""),
+    code = {
+      expect_error(
+        check_credentials("rato"),
+        class = "rata_no_credentials_set"
+      )
+    }
+  )
+
+  withr::with_envvar(
+    new = c("RATO_USER" = ""),
+    code = {
+      expect_error(
+        check_credentials("rato"),
+        class = "rata_no_credentials_set"
+      )
+    }
+  )
+
+  withr::with_envvar(
+    new = c("WFL_USER" = ""),
+    code = {
+      expect_error(
+        check_credentials("wfl"),
+        class = "rata_no_credentials_set"
+      )
+    }
+  )
+
+  withr::with_envvar(
+    new = c("WFL_PWD" = ""),
+    code = {
+      expect_error(
+        check_credentials("wfl"),
+        class = "rata_no_credentials_set"
+      )
+    }
+  )
+
+  withr::with_envvar(
+    new = c("WFL_USER" = "a_username",
+            "WFL_PWD" = "a_password",
+            "RATO_USER" = "",
+            "RATO_PWD" = ""),
+    code = {
+      expect_true(
+        check_credentials("wfl")
+      )
+
+      expect_error(
+        check_credentials("rato"),
+        class = "rata_no_credentials_set"
+      )
+    }
+  )
+})
+
+test_that("check_source() returns error on multiple sources", {
+  expect_error(
+    check_source(c("rato", "wfl")),
+    class = "rata_multiple_sources"
+  )
+})
+
+test_that("check_source() returns error on invalid source", {
+  expect_error(
+    check_source("invalid_source"),
+    class = "rlang_error"
+  )
+})
+
+test_that("check_source() returns error on missing source", {
+  expect_error(
+    check_source(),
+    class = "rata_no_source_specified"
+  )
+})